Updated Render
A new structure NameInterface and a new ConsoleRenderer function renderNameInterface for the user to input username are added. However, this contains only a basic UI of displaying user input. Rules of setting username and methods for saving user information and writing user information into files are currently not specified.
In main.cpp, after running all the render tasks in the vector, the a NameInterface object is created and renderNameInterface is called to show the rendering of the NameInterface.
Other minor changes (such as defining more macro that represent setting text colors and clearing the screen) are made to make the code neater.